Extract IGoogleSyncServiceRead and route Notifications meter off the full sync service
Reforge (vs TRUE pristine-HEAD baseline; the PROVIDED baseline-all.json is stale — it reports overall 57177 / GoogleIntegration 3523 from the IDENTICAL pristine HEAD that freshly re-scores to overall 59377 / GoogleIntegration 3667, a ~2200pt scoring drift unrelated to this change. To get an honest delta I stashed my work, regenerated baseline-true.json from pristine HEAD, then restored):

- GoogleIntegration: 3667 -> 3677 (+10); surface +10, internal +0
- Overall: 59377 -> 59381 (+4); surface +4, internal +0
- Outside target: 55710 -> 55704 (-6); improvement +6
- Weighted value: +2 (target -10 + 2x outside +6)

reforge_delta.py crashed (UnboundLocalError) AFTER printing the two section/overall lines: line 105 rebinds the `improvement` function name to a local whenever after_all has a `baseline` key — a pre-existing harness bug under .codex/skills (not editable). The two printed header lines are correct; outside/weighted computed by hand with the script's own formulas. Raw against the (stale) provided baseline-all.json was GI 3523->3677 (+154) / overall 57177->59381 (+2204), which is the stale-baseline artifact, NOT this change.

Architecture-review (score-blind 3-lens panel):
- Verdict: accept (unanimous)
- Lens A (relocation/internal): accept — Textbook application of the sanctioned section-read-write-split pattern (peters-hard-rules.md "IServiceRead interfaces for cross-section service calls"; memory/architecture/section-read-write-split.md; reference impl Teams). The two read methods and GoogleSyncOutboxEventSnapshot record were relocated into the new IGoogleSyncServiceRead, but IGoogleSyncService now inherits it, so the full surface is unchanged and no consumer lost a member — the relocation is the mechanism of the split, not a count-gaming hide. The genuine deletion is the narrowed cross-section dependency: verified via reforge that GetFailedSyncEventCountAsync's only production caller is NotificationMeterProvider (Notifications section), re-pointed to the 2-method read interface, while GetRecentOutboxEventsAsync's only caller is GoogleController (own-section), correctly kept on the full interface via inheritance. No accessibility narrowing — both interfaces are public; only the consumer's declared dependency type changed, and the real external caller is preserved and functional. The read methods cannot be inlined as LINQ over an existing DTO: they read google_sync_outbox_events, a table owned by GoogleIntegration that the Notifications section is forbidden to read directly (design-rules §2c), and no canonical read DTO over outbox events exists for the consumer. Behavior/contract preserved: implementation classes unchanged, DI registers the read interface from the same-scope full service in both credential branches, orphaned NodaTime using removed only because the record moved. Build SUCCESS, 430 tests pass, architecture/baseline tests green.
- Lens B (read-model): accept — Textbook application of the sanctioned section-read-write-split pattern (peters-hard-rules IServiceRead; memory/architecture/section-read-write-split.md; thesis C1, the section's #1 candidate). The two read methods were RELOCATED down to a new narrower IGoogleSyncServiceRead, with IGoogleSyncService inheriting them — no net new interface surface; the section's method count is unchanged. From a read-model-purist view: neither method can be inlined as caller LINQ. The thesis is explicit that this section has NO canonical cached <Section>Info read model; both methods query the section-owned google_sync_outbox table through the repository (CountFailedAsync / GetRecentAsync), and the Notifications section is forbidden from reading that table directly (design-rules §2c, hard-rule: only the owning repository touches its tables). There is no DTO to LINQ over, so these are the genuine minimum cross-section read surface, not a gameable scalar/predicate bag. GoogleSyncOutboxEventSnapshot is not a new DTO — it is the existing section-owned read projection moved alongside its method (the documented 'return *Info/snapshot, never EF entities' shape). Contract preserved: same impl classes satisfy both interfaces via inheritance (no impl change), DI registers the read interface resolving from the same-scope full service in both credential branches matching the Shifts/Teams convention, GoogleController correctly retains the full interface (calls mutations + inherited GetRecentOutboxEventsAsync), and the test mock was narrowed. Build green, architecture/baseline tests green. The genuinely deleted concept is the cross-section dependency EDGE on the mutation surface — real, nameable in one sentence, and exactly the ACCEPT-worthy 'consumer re-routed onto I<Section>ServiceRead' case.
- Lens C (behavior/contract): accept — Pure, correctly-executed cross-section read-write split. NotificationMeterProvider (the only cross-section consumer that injected the full 14-method mutation interface purely for a read) now depends on the 2-method IGoogleSyncServiceRead; verified it only ever called GetFailedSyncEventCountAsync. All other full-interface callers (GoogleController, Team controllers, infra jobs, GoogleAdminService) keep IGoogleSyncService and still resolve the inherited members. Implementations (GoogleWorkspaceSyncService, StubGoogleSyncService) unchanged and satisfy members via inheritance. DI registers the read interface resolving from the same-scope full service, so cache/transaction/audit/notification behavior is identical (same instance, no fork). The two read methods aggregate over the GoogleIntegration-owned google_sync_outbox_events table with error/processing state — unbounded, not a stable aggregate fact — so they correctly stay as service methods, not inline LINQ over an exposed DTO. No test removed or weakened: the only test change narrows a mock's declared type; all stubbings and assertions are identical and the failed-sync meter coverage is intact. Independently verified: build 0 errors / 140 pre-existing baseline warnings (none in touched/new files); tests 430 passed, 0 failed, 1 skipped, architecture/section baseline green. Pre-existing HUM0025 table-sharing on the outbox table is untouched and out of scope.
- Concept deleted: Notifications' compile-time dependency on the full 14-method IGoogleSyncService mutation surface — NotificationMeterProvider now depends only on the 2-method IGoogleSyncServiceRead contract.

Verification:
- Build: SUCCESS — 0 errors, 140 warnings (all pre-existing baseline HUM0025 table-sharing / HUM0014 warnings, none in the touched files or new file). Command: dotnet build Humans.slnx --disable-build-servers -v q
- Tests (dotnet test "H:/source/Humans/.worktrees/refactor-2026-05-29-googleintegration-workflow-3/tests/Humans.Application.Tests/Humans.Application.Tests.csproj" --filter "FullyQualifiedName~GoogleIntegration|FullyQualifiedName~Architecture|FullyQualifiedName~NotificationMeterProvider" -v q --disable-build-servers): PASSED — Failed: 0, Passed: 430, Skipped: 1, Total: 431. Architecture tests (section/baseline enforcement) all green, confirming no layering/section-isolation violation introduced.
- Reforge: GoogleIntegration after = 3677
